splitting peakfunctions module specific to fitpeak and wppf.
